+++ src/message.js
@@ -583,9 +583,9 @@ class Message extends Syncable {
    */
   _readAllBlobs(callback) {
     let count = 0;
-    const parts = this.parts.filter(part => Util.isBlob(part.body) && MessagePart.isTextualMimeType(part.mimeType));
+    const parts = this.parts.filter(part => Util.isBlob(part.body) && part.isTextualMimeType());
     parts.forEach((part) => {
-      part._fetchTextFromFile(part.body, (text) => {
+      part._fetchTextFromBlob(part.body, (text) => {
         part.body = text;
         count++;
         if (count === parts.length) callback();