+++ test/specs/unit/messages/messagePartSpec.js
@@ -89,7 +89,6 @@ describe("The MessageParts class", function() {
         it("Should initialize with an object", function() {
             expect(new layer.MessagePart({body: "hey"}).body).toEqual("hey");
             expect(new layer.MessagePart({mimeType: "text/hey"}).mimeType).toEqual("text/hey");
-            expect(new layer.MessagePart({encoding: "base64"}).encoding).toEqual("base64");
             expect(new layer.MessagePart({id: "Impart"}).id).toEqual("Impart");
         });
 
@@ -120,6 +119,13 @@ describe("The MessageParts class", function() {
             var text = new Array(layer.DbManager.MaxPartSize + 10).join('a');
             var blob = new Blob([text], {type : 'text/plain'});
             expect(new layer.MessagePart(blob).url).toEqual('');
+        });
+
+        it("Should convert the body to Blob if non-textual mimeType and non-blob body", function() {
+            expect(new layer.MessagePart({
+                body: "hey",
+                mimeType: "ho/hum"
+            }).body).toEqual(jasmine.any(Blob));
         });
 
         it("Should initialize with Content", function() {
@@ -207,7 +213,7 @@ describe("The MessageParts class", function() {
         });
     });
 
-    describe("The _fetchTextFromFile() method", function() {
+    describe("The _fetchTextFromBlob() method", function() {
         var part, message, blob, text;
         beforeEach(function() {
             text = new Array(layer.DbManager.MaxPartSize + 10).join('a');
@@ -218,13 +224,13 @@ describe("The MessageParts class", function() {
 
         it("Should return file if file is really a string", function() {
             var result;
-            part._fetchTextFromFile(text, function(data) { result = data;});
+            part._fetchTextFromBlob(text, function(data) { result = data;});
             expect(result).toEqual(text);
         });
 
         it("Should turn text blob to string", function(done) {
             var result;
-            part._fetchTextFromFile(text, function(data) {
+            part._fetchTextFromBlob(text, function(data) {
                 expect(data).toEqual(text);
                 done();
             });
@@ -236,7 +242,7 @@ describe("The MessageParts class", function() {
         var part, message, content;
         beforeEach(function() {
             content = new layer.Content({expiration: new Date()});
-            part = new layer.MessagePart({mimeType: "text/dog", _content: content});
+            part = new layer.MessagePart({mimeType: "food/dog", _content: content});
             message = conversation.createMessage({parts: [part]}).send();
             part.id = message.id + "/parts/0";
         });
@@ -261,17 +267,17 @@ describe("The MessageParts class", function() {
           expect(part._fetchContentComplete).toHaveBeenCalledWith(blob, undefined);
         });
 
-        it("Should call _fetchTextFromFile for text/plain", function() {
+        it("Should call _fetchTextFromBlob for text/plain", function() {
           var text = new Array(layer.DbManager.MaxPartSize + 10).join('a');
           var blob = new Blob([text], {type : 'text/plain'});
           part = new layer.MessagePart(blob);
-          spyOn(part, "_fetchTextFromFile").and.callFake(function(file, callback) {callback(text);});;
+          spyOn(part, "_fetchTextFromBlob").and.callFake(function(file, callback) {callback(text);});;
           spyOn(part, "_fetchContentComplete");
           var spy = jasmine.createSpy('callback');
           part._fetchContentCallback(null, blob, spy);
 
           // Posttest
-          expect(part._fetchTextFromFile).toHaveBeenCalledWith(blob, jasmine.any(Function));
+          expect(part._fetchTextFromBlob).toHaveBeenCalledWith(blob, jasmine.any(Function));
           expect(part._fetchContentComplete).toHaveBeenCalledWith(text, spy);
         });
 
@@ -345,7 +351,7 @@ describe("The MessageParts class", function() {
           part = layer.MessagePart._createFromServer({
               id: message.id + "/parts/3",
               body: "jane",
-              encoding: "john",
+              mime_type: 'dog/food',
               content: {
                   id: "jill",
                   download_url: "fred",
@@ -393,7 +399,7 @@ describe("The MessageParts class", function() {
           part = layer.MessagePart._createFromServer({
               id: "joe",
               body: "jane",
-              encoding: "john",
+              mime_type: 'dog/food',
               content: {
                   id: "jill",
                   download_url: "fred",
@@ -493,7 +499,7 @@ describe("The MessageParts class", function() {
         it("Should trigger with body and mime_type", function() {
             var part = new layer.MessagePart({
                 body: "hey",
-                mimeType: "ho"
+                mimeType: "text/ho"
             });
             spyOn(part, "trigger");
 
@@ -503,34 +509,15 @@ describe("The MessageParts class", function() {
             // Posttest
             expect(part.trigger).toHaveBeenCalledWith("parts:send", {
                 body: "hey",
-                mime_type: "ho"
+                mime_type: "text/ho"
             });
         });
 
-        it("Should trigger with encoding", function() {
-            var part = new layer.MessagePart({
-                body: "hey",
-                mimeType: "ho",
-                encoding: "fred"
-            });
-            spyOn(part, "trigger");
-
-            // Run
-            part._sendBody();
-
-            // Posttest
-            expect(part.trigger).toHaveBeenCalledWith("parts:send", {
-                body: "hey",
-                mime_type: "ho",
-                encoding: "fred"
-            });
-        });
 
         it("Should throw error on non-string", function() {
            var part = new layer.MessagePart({
                 body: {hey: "ho"},
-                mimeType: "ho",
-                encoding: "fred"
+                mimeType: "text/ho"
             });
 
             // Run
@@ -569,12 +556,18 @@ describe("The MessageParts class", function() {
     describe("The _sendBlob() method", function() {
         it("Should send small blobs", function(done) {
             var part = new layer.MessagePart({
-                body: new Blob([atob("abc")], {type: "text/plain"}),
+                body: new Blob([atob("abc")], {type: "fred"}),
                 mimeType: "fred"
             });
 
-            spyOn(part, "_sendBody").and.callFake(function() {
-                expect(true).toBe(true);
+            spyOn(part, "trigger").and.callFake(function(eventName, data) {
+                expect(eventName).toEqual("parts:send");
+                expect(data).toEqual({
+                    encoding: "base64",
+                    mime_type: "fred",
+                    body: jasmine.any(String)
+                });
+                expect(data.body.length > 0).toBe(true);
                 done();
             });
 
@@ -807,6 +800,7 @@ describe("The MessageParts class", function() {
                 body: "hey",
                 mimeType: "text/plain2"
             });
+            spyOn(part, "isTextualMimeType").and.returnValue(false);
             expect(part.getText()).toEqual("");
         });
     });
@@ -835,7 +829,7 @@ describe("The MessageParts class", function() {
             part = layer.MessagePart._createFromServer({
                 id: "joe",
                 body: "jane",
-                encoding: "john",
+                mime_type: 'text/plain',
                 content: {
                     id: "jill"
                 }
@@ -854,10 +848,6 @@ describe("The MessageParts class", function() {
             expect(part.body).toEqual("jane");
         });
 
-        it("Should have a correct encoding", function() {
-            expect(part.encoding).toEqual("john");
-        });
-
         it("Should have a correct content", function() {
             expect(part._content instanceof layer.Content).toBe(true);
             expect(part._content.id).toEqual("jill");
@@ -871,6 +861,7 @@ describe("The MessageParts class", function() {
           part = layer.MessagePart._createFromServer({
                 id: "joe",
                 body: "jane",
+                mime_type: "text/plain",
                 encoding: "john"
             });
             expect(part.hasContent).toEqual(false);
@@ -884,6 +875,7 @@ describe("The MessageParts class", function() {
               id: "joe",
               body: "jane",
               encoding: "john",
+              mime_type: "text/plain",
               content: {
                   id: "jill",
                   download_url: "fred",